requirements.txt
fastapi
uvicorn[standard]
python-dotenv
twilio
websockets
openai
deepgram-sdk
requests
audioop-lts
httpx
